Fixed a bug (and other potential sources of bugs) due to overflows of integers when not using castings to size_t. Thanks to Maksim Struchalin for pointing it out.
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/OmicABEL@1898 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ src/databel.c
@@ -39,7 +39,7 @@ struct databel_fvi * load_databel_fvi( char *path )
 	// Header
 	fread( &fvi->fvi_header, sizeof(databel_fvi_header), 1, f );
 	// Labels
-	data_size = (fvi->fvi_header.numVariables +fvi->fvi_header.numObservations ) * 
+	data_size = ((size_t)fvi->fvi_header.numVariables + fvi->fvi_header.numObservations ) * 
 		         fvi->fvi_header.namelength * sizeof(char);
 	fvi->fvi_data = (char *) fgls_malloc ( data_size );
 	// Load labels
